refactor: extract group info component
GroupInfo component is also used in new topic page.
