+++ packages/website/src/pages/index/group/topic/[id]/index.module.less
@@ -15,62 +15,3 @@
     margin-left: 12px;
   }
 }
-
-// Right Col
-.groupInfo {
-  display: flex;
-}
-
-.groupDetails {
-  margin-left: 10px;
-
-  > a {
-    display: block;
-    margin-bottom: 6px;
-    font-weight: 600;
-    font-size: 16px;
-    line-height: 22px;
-  }
-
-  > span {
-    font-size: 12px;
-    line-height: 17px;
-    color: @gray-60;
-  }
-}
-
-.groupDescription {
-  margin-top: 20px;
-  font-size: 14px;
-  line-height: 20px;
-  color: @gray-80;
-}
-
-.groupOpinions {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 20px;
-
-  > :global(.bgm-button) {
-    border: 2px solid @gray-10;
-    border-radius: 15px;
-    height: 30px;
-    line-height: 24px;
-    padding: 0 32px;
-
-    a {
-      font-weight: 600;
-      color: @gray-60;
-      text-decoration: none;
-    }
-
-    &:hover {
-      background-color: @gray-10;
-
-      a {
-        color: @gray-80;
-      }
-    }
-  }
-}
